backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
langchain==0.3.7
langchain-groq==0.2.1
langchain-core==0.3.15
langgraph==0.2.45
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
httpx==0.27.2
python-dotenv==1.0.1